pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibesafe"
version = "0.1.0"
description = "Standardized project management practices for consistent development"
readme = "README.md"
authors = [
    {name = "Neil Lawrence", email = "example@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.6"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.6"
warn_return_any = true
warn_unused_configs = true 